Fix lint

+++ tooling/replayer/src/main.rs
@@ -359,7 +359,7 @@ async fn replay_latest_block(
         CacheLevel::Off => {
             // We don't want any cache
             tracing::info!("Deleting cache: Caching is disabled");
-            delete_cache(network, latest_block as u64);
+            delete_cache(network, latest_block);
         }
         CacheLevel::Failed => {
             // We only want caches that failed
@@ -367,7 +367,7 @@ async fn replay_latest_block(
                 tracing::info!(
                     "Deleting cache: Execution was successful and Cache Level is 'failed'"
                 );
-                delete_cache(network, latest_block as u64);
+                delete_cache(network, latest_block);
             } else {
                 // I prefer to be explicit about keeping the cache file
                 tracing::info!(
